+++ Assets/Scripts/Level Section Builder/Section_Obstacle_Object.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Numerics;
 using UnityEngine;
 using EditorAttributes;
 using UnityEngine.Serialization;
@@ -21,8 +20,14 @@ public class Section_Obstacle_Object : SectionBuilderObject
         [FormerlySerializedAs("overridedData")] 
         [ShowField(nameof(overrideData))] public ObstacleData overriddenData = new();
 
+        // //TODO: Remember to delete these two variables
         // Pipe Data
-        [ShowField(nameof(obstacleType), ObstacleType.SewerPipe)] public Pipe_Obstacle_Data pipeObstacleData = new();
+        [ShowField(nameof(obstacleType), ObstacleType.SewerPipe)]
+        public NewPipeObstacleData pipeObstacleData = new();
+        
+        // Bridge Data
+        [ShowField(nameof(obstacleType), ObstacleType.Bridge)]
+        public ObstacleBridgeData obstacleBridgeData = new();
     }
 
     /// <summary> The data of the obstacle shared with the Game_Section_Manager and Section_Builder </summary>
@@ -31,7 +36,7 @@ public class Section_Obstacle_Object : SectionBuilderObject
 
     public enum ObstacleType
     {
-        TrashPile, WideTrashPile, SewerPipe
+        TrashPile, WideTrashPile, SewerPipe, Bridge
     }
 
     public override void Register(SectionContentBuilder section)
@@ -43,27 +48,6 @@ public class Section_Obstacle_Object : SectionBuilderObject
     protected override void AdditionalDebug()
     {
         DrawItem(Color.cyan, Vector3.one, transform.position);
-        name = new($"{ObjectType.Obstacle} - {sectionData.obstacleType}");
-        
-        if (sectionData.obstacleType != ObstacleType.SewerPipe) return;
-        
-        // Do Sewer Pipe specific Visual Debugging
-        
-        // TODO: Draw boxes displaying the connection of pipes
-        var data = sectionData.pipeObstacleData;
-        for (int i = 0; i < data.amount; i++)
-        {
-            var direction = data.pipeConnection switch
-            {
-                Pipe_Obstacle_Data.PipeConnection.Left => Vector3.left,
-                Pipe_Obstacle_Data.PipeConnection.Top => Vector3.up,
-                Pipe_Obstacle_Data.PipeConnection.Bottom => Vector3.down,
-                Pipe_Obstacle_Data.PipeConnection.Right => Vector3.right,
-                _ => Vector3.zero
-            };
-            var position = transform.position + i * (direction * data.distancePerPipe);
-            DrawItem(Color.cyan, Vector3.one, position);
-        }
     }
 
     protected override void AdditionalDebugSelected()